Created client debug UI, and fixed rotations being sent over network even if there was no change

+++ Assets/Scripts/Client.cs
@@ -93,7 +93,7 @@ public class Client
         _client.Dispose();
     }
 
-    public float GetRTT()
+    public uint GetRTT()
     {
         return _peer.RoundTripTime;
     }
